+++ index.test-d.ts
@@ -29,6 +29,7 @@ import insert, {
   onConflictThrowError,
   onConflictUseGivenValue,
   onNotFoundDoNothing,
+  onNotFoundThrowError,
   onNotFoundUseGivenValue,
 } from './insert';
 import type {
@@ -62,6 +63,7 @@ declare function cmpNumberAndString(a: number, b: string): number;
 expectType<types.ImmutableTree<string>>(create<string>(''));
 expectType<types.ImmutableTree<number | null> | null>(insertByKey<number | null, number>(numberTree, 0, cmpNullableNumbers, onConflictKeepTreeValue, onNotFoundUseGivenValue));
 expectType<types.ImmutableTree<string> | null>(insertByKey<string, number>(stringTree, 0, cmpNumberAndString, onConflictKeepTreeValue, onNotFoundDoNothing));
+expectType<types.ImmutableTree<string> | null>(insertByKey<string, number>(stringTree, 0, cmpNumberAndString, onConflictKeepTreeValue, onNotFoundThrowError));
 expectType<types.ImmutableTree<string>>(insert<string>(stringTree, '', cmpStrings));
 expectType<types.ImmutableTree<string>>(insert<string>(stringTree, '', cmpStrings, NOOP));
 expectType<types.ImmutableTree<string>>(insert<string>(stringTree, '', cmpStrings, REPLACE));
@@ -142,6 +144,7 @@ expectAssignable<InsertConflictHandler<unknown, unknown>>(onConflictUseGivenValu
 // InsertNotFoundHandler
 expectType<InsertNotFoundHandler<string, number>>((a: number) => String(a));
 expectAssignable<InsertNotFoundHandler<unknown, unknown>>(onNotFoundDoNothing);
+expectAssignable<InsertNotFoundHandler<unknown, unknown>>(onNotFoundThrowError);
 expectAssignable<InsertNotFoundHandler<unknown, unknown>>(onNotFoundUseGivenValue);
 
 // Wrong 'onNotFound' function type.